+++ inst/r-bindings/package.json
@@ -6,6 +6,7 @@
   "dependencies": {
     "util-functions": "*",
     "tsconfig": "*",
+    "python-ts-parser": "*",
     "typescript": "^5.0.4"
   },
   "devDependencies": {
